services/gateway/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
boto3==1.37.28
aiokafka==0.12.0
python-dotenv==1.1.0